requirements.txt
azure-functions==1.19.0
python-dotenv
requests
pyodbc
pandas
requests==2.31.0
numpy